Secure edit, update, and destroy actions, add controller tests for this

+++ app/controllers/grams_controller.rb
@@ -1,12 +1,12 @@
 class GramsController < ApplicationController
-  before_action :authenticate_user!, only: [:new, :create]
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
   
   def index
   end
   
   def show
     @gram = Gram.find_by_id(params[:id])
-    render_not_found if @gram.blank?
+    return render_not_found if @gram.blank?
   end
   
   def new
@@ -26,12 +26,18 @@ class GramsController < ApplicationController
   
   def edit
     @gram = Gram.find_by_id(params[:id])
-    render_not_found if @gram.blank?
+    return render_not_found if @gram.blank?
+    return render_not_found(:forbidden) if @gram.user != current_user
   end
   
   def update
     @gram = Gram.find_by_id(params[:id])
     return render_not_found if @gram.blank?
+    return render_not_found(:forbidden) if @gram.user != current_user
+    
+    if @gram.user != current_user
+      return render text: 'Forbidden :(', status: :forbidden
+    end
     
     if @gram.update_attributes(gram_params)
       redirect_to root_path
@@ -44,6 +50,7 @@ class GramsController < ApplicationController
   def destroy
     @gram = Gram.find_by_id(params[:id])
     return render_not_found if @gram.blank?
+    return render_not_found(:forbidden) if @gram.user != current_user
     
     if @gram.destroy
       redirect_to root_path
@@ -59,8 +66,8 @@ class GramsController < ApplicationController
       params.require(:gram).permit(:message)
     end
     
-    def render_not_found
-      render text: 'Not Found :(', status: :not_found
+    def render_not_found(status = :not_found)
+      render text: '#{status.to_s.titleize} :(', status: status
     end
   
 end
